CSC_580/module_06/requirements.txt
tensorflow[and-cuda]
numpy
seaborn
pillow
scikit-learn